requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
annotated-types==0.5.0
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
cachetools==5.3.2
certifi==2023.7.22
cffi==1.16.0
cfgv==3.4.0
charset-normalizer==3.2.0
click==8.1.7
cryptography==41.0.7
distlib==0.3.7
Django==4.2.5
django-cors-headers==4.2.0
djangogrpcframework==0.2.1
djangorestframework==3.14.0
djangorestframework-api-key==3.0.0
djangorestframework-simplejwt==5.3.0
docker==6.1.3
drf-spectacular==0.26.5
filelock==3.12.4
flake8==4.0.1
flake8-docstrings==1.6.0
frozenlist==1.4.0
google-ai-generativelanguage==0.4.0
google-api-core==2.15.0
google-auth==2.25.2
google-generativeai==0.3.2
googleapis-common-protos==1.62.0
grpcio==1.60.0
grpcio-status==1.60.0
identify==2.5.29
idna==3.4
inflection==0.5.1
isort==5.10.1
json-fix==0.5.2
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
lazydocs==0.4.8
mccabe==0.6.1
multidict==6.0.4
nodeenv==1.8.0
openai==0.28.0
packaging==23.1
platformdirs==3.10.0
pre-commit==2.17.0
proto-plus==1.23.0
protobuf==4.25.1
psycopg2==2.9.7
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycodestyle==2.8.0
pycparser==2.21
pydantic==2.3.0
pydantic-settings==2.0.3
pydantic_core==2.6.3
pydocstyle==6.3.0
pyflakes==2.4.0
PyJWT==2.8.0
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.30.2
requests==2.31.0
RestrictedPython==6.2
rpds-py==0.10.3
rsa==4.9
snowballstemmer==2.2.0
sqlparse==0.4.4
toml==0.10.2
tqdm==4.66.1
typer==0.9.0
typing_extensions==4.8.0
uritemplate==4.1.1
urllib3==2.0.5
virtualenv==20.24.5
websocket-client==1.7.0
yarl==1.9.2
